+++ src/test/java/org/reaktivity/specification/mqtt/streams/PublishIT.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-package org.reaktivity.specification.mqtt;
+package org.reaktivity.specification.mqtt.streams;
 
 import static java.util.concurrent.TimeUnit.SECONDS;
 import static org.junit.rules.RuleChain.outerRule;
@@ -27,10 +27,10 @@ import org.kaazing.k3po.junit.annotation.ScriptProperty;
 import org.kaazing.k3po.junit.annotation.Specification;
 import org.kaazing.k3po.junit.rules.K3poRule;
 
-public class SubscribeIT
+public class PublishIT
 {
     private final K3poRule k3po = new K3poRule()
-            .addScriptRoot("scripts", "org/reaktivity/specification/mqtt/subscribe");
+            .addScriptRoot("scripts", "org/reaktivity/specification/mqtt/publish");
     private final TestRule timeout = new DisableOnDebug(new Timeout(5, SECONDS));
 
     @Rule
@@ -38,10 +38,10 @@ public class SubscribeIT
 
     @Test
     @Specification({
-            "${scripts}/client",
-            "${scripts}/server"})
+            "${scripts}/at.most.once/client",
+            "${scripts}/at.most.once/server"})
     @ScriptProperty("serverTransport \"nukleus://streams/mqtt#0\"")
-    public void subscribe() throws Exception
+    public void publishAtMostOnce() throws Exception
     {
         k3po.start();
         k3po.notifyBarrier("ROUTED_SERVER");
